feat: move AppConfig to a separate module
